Support using typedef aliases or the original classes interchangeably.
Injector sometimes uses InjectorMacros.getType() to get a type name,
and other times uses Type.getClassName().  This causes problems when
you use a typedef that is an alias to a class - the macro name and
the runtime name are different.

This fix uses `Context.follow()` to get the original class behind any
typedef aliases, so that the macro name matches the runtime name.

Unit tests included.

On a side note, this also fixed a different test that was failing in V2.
The error before was:

    Value [null] was not of type: minject.support.injectees.childinjectors.RobotToes at minject.ChildInjectorTest#childInjectorUsesParentInjectorForMissingRules (105)

With this commit, that test passes as expected.