Fix: childInjector.getResponse( something_in_parent_injector ) created blank rule.
As a result,

	parentInjector.mapValue( String, "Hello" );
	childInjector.getResponse( String );
	childInjector.getInstance( String ); // No rule defined for class "String" named "null"

I've created some unit tests.
Please note they use @:access(minject.Injector) so I can check if a rule exists
ONLY on the parent injector, not the child injector. Merely testing the code above
did not guarantee that the rule (with a null result) was not created on the child
injector.